chore: move experimental face-target to main
