--- src/experimental/face-target/aframe.ts
@@ -1,5 +1,0 @@
-import './aframe/face';
-import './aframe/face-default-occluder';
-import './aframe/face-occluder';
-import './aframe/face-target';
-import './aframe/face-system';
