--- src/experimental/face-target/aframe/face-system.ts
@@ -1,242 +1,0 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import { Scene } from 'aframe';
-import { UI } from '../../../ui/ui';
-import Controller from '../controller';
-import { Helper } from '../../../libs';
-import { AR_COMPONENT_NAME, AR_EVENT_NAME } from '../utils/constant';
-import { AR_STATE, AR_ELEMENT_TAG, GLOBAL_AR_EVENT_NAME } from '../../../utils/constant';
-import screenResizer from '../../../utils/screen-resizer';
-
-const { Controller: ControllerClass, UI: UIClass } = window.MINDAR.FACE;
-
-AFRAME.registerSystem(AR_COMPONENT_NAME.FACE_SYSTEM, {
-  container: Helper.castTo<HTMLDivElement>(null),
-  video: Helper.castTo<HTMLVideoElement>(null),
-  anchorEntities: [] as any[],
-  faceMeshEntities: [] as any[],
-  filterMinCF: -Infinity,
-  filterBeta: Infinity,
-  controller: Helper.castTo<Controller>(null),
-  ui: Helper.castTo<UI>(null),
-  el: null as any,
-  shouldFaceUser: true,
-  lastHasFace: false,
-
-  init: function () {
-    this.anchorEntities = [];
-    this.faceMeshEntities = [];
-  },
-
-  setup: function ({
-    uiLoading,
-    uiScanning,
-    uiError,
-    filterMinCF,
-    filterBeta,
-    shouldFaceUser,
-  }: {
-    uiLoading: string;
-    uiScanning: string;
-    uiError: string;
-    filterMinCF: number;
-    filterBeta: number;
-    shouldFaceUser: boolean;
-  }) {
-    this.ui = new UIClass({ uiLoading, uiScanning, uiError });
-    this.filterMinCF = filterMinCF;
-    this.filterBeta = filterBeta;
-    this.shouldFaceUser = shouldFaceUser;
-  },
-
-  registerFaceMesh: function (el: any) {
-    this.faceMeshEntities.push({ el });
-  },
-
-  registerAnchor: function (el: any, targetIndex: number) {
-    this.anchorEntities.push({ el, targetIndex });
-  },
-
-  start: function () {
-    this.ui.showLoading();
-
-    this.container = this.el.sceneEl.parentNode;
-
-    this._startVideo();
-  },
-
-  stop: function () {
-    this.pause();
-
-    if (!this.video) return;
-
-    const { srcObject } = this.video;
-
-    if (!srcObject) return;
-
-    const tracks = (srcObject as MediaStream).getTracks();
-
-    tracks.forEach(function (track) {
-      track.stop();
-    });
-
-    this.video.remove();
-  },
-
-  switchCamera: function () {
-    this.shouldFaceUser = !this.shouldFaceUser;
-
-    this.stop();
-    this.start();
-  },
-
-  pause: function (keepVideo = false) {
-    if (!keepVideo) this.video.pause();
-
-    this.controller.stopProcessVideo();
-  },
-
-  unpause: function () {
-    this.video.play();
-    this.controller.processVideo(this.video);
-  },
-
-  _startVideo: async function () {
-    this.video = Helper.castTo<HTMLVideoElement>(document.createElement('video'));
-
-    this.video.setAttribute('autoplay', '');
-    this.video.setAttribute('muted', '');
-    this.video.setAttribute('playsinline', '');
-
-    this.video.style.position = 'absolute';
-    this.video.style.top = '0px';
-    this.video.style.left = '0px';
-    this.video.style.zIndex = '-2';
-
-    this.container.appendChild(this.video);
-
-    if (!navigator.mediaDevices || !navigator.mediaDevices.getUserMedia) {
-      // TODO: show unsupported error
-      this.el.emit(AR_STATE.AR_ERROR, { error: 'VIDEO_FAIL' });
-      this.ui.showCompatibility();
-
-      return;
-    }
-
-    try {
-      const DEVICES = await navigator.mediaDevices.enumerateDevices();
-      const devices = DEVICES.filter((device) => device.kind === 'videoinput');
-
-      let facingMode: VideoFacingModeEnum = 'environment';
-
-      if (devices.length > 1) facingMode = this.shouldFaceUser ? 'user' : 'environment';
-
-      const stream = await navigator.mediaDevices.getUserMedia({
-        audio: false,
-        video: {
-          facingMode,
-        },
-      });
-
-      this.video.addEventListener(GLOBAL_AR_EVENT_NAME.LOADED_METADATA, async () => {
-        this.video.setAttribute('width', this.video.videoWidth.toString());
-        this.video.setAttribute('height', this.video.videoHeight.toString());
-
-        await this._setupAR();
-
-        this._processVideo();
-        this.ui.hideLoading();
-      });
-
-      this.video.srcObject = stream;
-    } catch (err) {
-      console.log('getUserMedia error', err);
-      this.el.emit(AR_STATE.AR_ERROR, { error: 'VIDEO_FAIL' });
-    }
-  },
-
-  _processVideo: function () {
-    this.controller.onUpdate = ({ hasFace, estimateResult }) => {
-      if (hasFace && !this.lastHasFace) {
-        this.el.emit(AR_EVENT_NAME.TARGET_FOUND);
-      }
-
-      if (!hasFace && this.lastHasFace) {
-        this.el.emit(AR_EVENT_NAME.TARGET_LOST);
-      }
-
-      this.lastHasFace = !!hasFace;
-
-      if (hasFace && estimateResult) {
-        const { faceMatrix } = estimateResult;
-        for (let i = 0; i < this.anchorEntities.length; i++) {
-          const landmarkMatrix = this.controller.getLandmarkMatrix(
-            this.anchorEntities[i].targetIndex
-          );
-
-          this.anchorEntities[i].el.updateVisibility(true);
-          this.anchorEntities[i].el.updateMatrix(landmarkMatrix);
-        }
-
-        for (let i = 0; i < this.faceMeshEntities.length; i++) {
-          this.faceMeshEntities[i].el.updateVisibility(true);
-          this.faceMeshEntities[i].el.updateMatrix(faceMatrix);
-        }
-      } else {
-        for (let i = 0; i < this.anchorEntities.length; i++) {
-          this.anchorEntities[i].el.updateVisibility(false);
-        }
-
-        for (let i = 0; i < this.faceMeshEntities.length; i++) {
-          this.faceMeshEntities[i].el.updateVisibility(false);
-        }
-      }
-    };
-
-    this.controller.processVideo(this.video);
-  },
-
-  _setupAR: async function () {
-    this.controller = new ControllerClass({
-      filterMinCF: this.filterMinCF,
-      filterBeta: this.filterBeta,
-    }) as any;
-
-    this._resize();
-
-    await this.controller.setup(this.video);
-    await this.controller.dummyRun(this.video);
-
-    const { fov, aspect, near, far } = this.controller.getCameraParams();
-
-    const camera = new AFRAME.THREE.PerspectiveCamera() as any;
-    camera.fov = fov;
-    camera.aspect = aspect;
-    camera.near = near;
-    camera.far = far;
-    camera.updateProjectionMatrix();
-
-    const cameraEle = this.container.getElementsByTagName(AR_ELEMENT_TAG.A_CAMERA)[0] as any;
-    cameraEle.setObject3D(AR_ELEMENT_TAG.CAMERA, camera);
-    cameraEle.setAttribute(AR_ELEMENT_TAG.CAMERA, 'active', true);
-
-    for (let i = 0; i < this.faceMeshEntities.length; i++)
-      this.faceMeshEntities[i].el.addFaceMesh(
-        this.controller.createThreeFaceGeometry(AFRAME.THREE)
-      );
-
-    this._resize();
-    window.addEventListener(GLOBAL_AR_EVENT_NAME.SCREEN_RESIZE, this._resize.bind(this));
-    this.el.emit(AR_STATE.AR_READY);
-  },
-
-  _resize: function () {
-    screenResizer(this.video, this.container);
-
-    const sceneEl = this.container.getElementsByTagName(AR_ELEMENT_TAG.A_SCENE)[0] as Scene;
-
-    sceneEl.style.top = this.video.style.top;
-    sceneEl.style.left = this.video.style.left;
-    sceneEl.style.width = this.video.style.width;
-    sceneEl.style.height = this.video.style.height;
-  },
-});
