--- src/experimental/face-target/face-geometry/face-data/index.ts
@@ -1,8 +1,0 @@
-//This is a generated file
-
-import faces from './faces';
-import landmarkBasis from './landmark-basis';
-import positions from './positions';
-import uvs from './uvs';
-
-export { faces, landmarkBasis, positions, uvs };
