--- src/experimental/face-target/face-geometry/face-data/uvs.ts
@@ -1,474 +1,0 @@
-//This is a generated file
-
-const uvs = [
-  [0.499977, 0.347466],
-  [0.500026, 0.452513],
-  [0.499974, 0.397628],
-  [0.482113, 0.528021],
-  [0.500151, 0.472844],
-  [0.49991, 0.501747],
-  [0.499523, 0.598938],
-  [0.289712, 0.619236],
-  [0.499955, 0.687602],
-  [0.499987, 0.730081],
-  [0.500023, 0.89295],
-  [0.500023, 0.333766],
-  [0.500016, 0.320776],
-  [0.500023, 0.307652],
-  [0.499977, 0.304722],
-  [0.499977, 0.294066],
-  [0.499977, 0.280615],
-  [0.499977, 0.262981],
-  [0.499968, 0.218629],
-  [0.499816, 0.437019],
-  [0.473773, 0.42609],
-  [0.104907, 0.745859],
-  [0.36593, 0.590424],
-  [0.338758, 0.586975],
-  [0.31112, 0.59054],
-  [0.274658, 0.610869],
-  [0.393362, 0.596294],
-  [0.345234, 0.655989],
-  [0.370094, 0.653924],
-  [0.319322, 0.652735],
-  [0.297903, 0.646409],
-  [0.247792, 0.58919],
-  [0.396889, 0.157245],
-  [0.280098, 0.6244],
-  [0.10631, 0.600044],
-  [0.209925, 0.608647],
-  [0.355808, 0.465594],
-  [0.471751, 0.349596],
-  [0.474155, 0.319808],
-  [0.439785, 0.342771],
-  [0.414617, 0.333459],
-  [0.450374, 0.319139],
-  [0.428771, 0.317309],
-  [0.374971, 0.272195],
-  [0.486717, 0.452371],
-  [0.485301, 0.472605],
-  [0.257765, 0.68551],
-  [0.401223, 0.544828],
-  [0.429819, 0.451385],
-  [0.421352, 0.466259],
-  [0.276896, 0.467943],
-  [0.48337, 0.500413],
-  [0.337212, 0.717117],
-  [0.296392, 0.706757],
-  [0.169295, 0.806186],
-  [0.44758, 0.69739],
-  [0.39239, 0.646112],
-  [0.35449, 0.303216],
-  [0.067305, 0.269895],
-  [0.442739, 0.427174],
-  [0.457098, 0.415208],
-  [0.381974, 0.305289],
-  [0.392389, 0.305797],
-  [0.277076, 0.728068],
-  [0.422552, 0.436767],
-  [0.385919, 0.718636],
-  [0.383103, 0.74416],
-  [0.331431, 0.880286],
-  [0.229924, 0.767997],
-  [0.364501, 0.810886],
-  [0.229622, 0.700459],
-  [0.173287, 0.721252],
-  [0.472879, 0.333802],
-  [0.446828, 0.331473],
-  [0.422762, 0.32611],
-  [0.445308, 0.419934],
-  [0.388103, 0.306039],
-  [0.403039, 0.29346],
-  [0.403629, 0.306047],
-  [0.460042, 0.442861],
-  [0.431158, 0.307634],
-  [0.452182, 0.307634],
-  [0.475387, 0.307634],
-  [0.465828, 0.22081],
-  [0.472329, 0.263774],
-  [0.473087, 0.282143],
-  [0.473122, 0.295374],
-  [0.473033, 0.304722],
-  [0.427942, 0.304722],
-  [0.426479, 0.29646],
-  [0.423162, 0.288154],
-  [0.418309, 0.279937],
-  [0.390095, 0.360427],
-  [0.013954, 0.439966],
-  [0.499914, 0.419853],
-  [0.4132, 0.3046],
-  [0.409626, 0.298177],
-  [0.46808, 0.398465],
-  [0.422729, 0.414015],
-  [0.46308, 0.406216],
-  [0.37212, 0.526586],
-  [0.334562, 0.503927],
-  [0.411671, 0.453035],
-  [0.242176, 0.852324],
-  [0.290777, 0.798554],
-  [0.327338, 0.743473],
-  [0.39951, 0.251079],
-  [0.441728, 0.738324],
-  [0.429765, 0.812166],
-  [0.412198, 0.891099],
-  [0.288955, 0.601048],
-  [0.218937, 0.564589],
-  [0.412782, 0.60103],
-  [0.257135, 0.64456],
-  [0.427685, 0.562039],
-  [0.44834, 0.463064],
-  [0.17856, 0.542446],
-  [0.247308, 0.542806],
-  [0.286267, 0.532325],
-  [0.332828, 0.539288],
-  [0.368756, 0.552793],
-  [0.398964, 0.567345],
-  [0.47641, 0.594194],
-  [0.189241, 0.476076],
-  [0.228962, 0.651049],
-  [0.490726, 0.437599],
-  [0.40467, 0.514867],
-  [0.019469, 0.598436],
-  [0.426243, 0.579569],
-  [0.396993, 0.451203],
-  [0.26647, 0.623023],
-  [0.439121, 0.481042],
-  [0.032314, 0.355643],
-  [0.419054, 0.612845],
-  [0.462783, 0.494253],
-  [0.238979, 0.220255],
-  [0.198221, 0.168062],
-  [0.10755, 0.459245],
-  [0.18361, 0.259743],
-  [0.13441, 0.666317],
-  [0.385764, 0.116846],
-  [0.490967, 0.420622],
-  [0.382385, 0.491427],
-  [0.174399, 0.602329],
-  [0.318785, 0.603765],
-  [0.343364, 0.599403],
-  [0.3961, 0.289783],
-  [0.187885, 0.411462],
-  [0.430987, 0.055935],
-  [0.318993, 0.101715],
-  [0.266248, 0.130299],
-  [0.500023, 0.809424],
-  [0.499977, 0.045547],
-  [0.36617, 0.601178],
-  [0.393207, 0.604463],
-  [0.410373, 0.60892],
-  [0.194993, 0.657898],
-  [0.388665, 0.637716],
-  [0.365962, 0.644029],
-  [0.343364, 0.644643],
-  [0.318785, 0.64166],
-  [0.301415, 0.636844],
-  [0.058133, 0.680924],
-  [0.301415, 0.612551],
-  [0.499988, 0.381566],
-  [0.415838, 0.375804],
-  [0.445682, 0.433923],
-  [0.465844, 0.379359],
-  [0.499923, 0.648476],
-  [0.288719, 0.180054],
-  [0.335279, 0.14718],
-  [0.440512, 0.097581],
-  [0.128294, 0.208059],
-  [0.408772, 0.626106],
-  [0.455607, 0.548199],
-  [0.499877, 0.09101],
-  [0.375437, 0.075808],
-  [0.11421, 0.384978],
-  [0.448662, 0.304722],
-  [0.44802, 0.295368],
-  [0.447112, 0.284192],
-  [0.444832, 0.269206],
-  [0.430012, 0.233191],
-  [0.406787, 0.314327],
-  [0.400738, 0.318931],
-  [0.3924, 0.322297],
-  [0.367856, 0.336081],
-  [0.247923, 0.398667],
-  [0.45277, 0.57915],
-  [0.436392, 0.640113],
-  [0.416164, 0.631286],
-  [0.413386, 0.307634],
-  [0.228018, 0.316428],
-  [0.468268, 0.647329],
-  [0.411362, 0.195673],
-  [0.499989, 0.530175],
-  [0.479154, 0.557346],
-  [0.499974, 0.560363],
-  [0.432112, 0.506411],
-  [0.499886, 0.133083],
-  [0.499913, 0.178271],
-  [0.456549, 0.180799],
-  [0.344549, 0.254561],
-  [0.378909, 0.42599],
-  [0.374293, 0.219815],
-  [0.319688, 0.429262],
-  [0.357155, 0.39573],
-  [0.295284, 0.378419],
-  [0.44775, 0.137523],
-  [0.410986, 0.491277],
-  [0.313951, 0.224692],
-  [0.354128, 0.187447],
-  [0.324548, 0.296007],
-  [0.189096, 0.3537],
-  [0.279777, 0.285342],
-  [0.133823, 0.317299],
-  [0.336768, 0.355267],
-  [0.429884, 0.533478],
-  [0.455528, 0.451377],
-  [0.437114, 0.441104],
-  [0.467288, 0.470075],
-  [0.414712, 0.66478],
-  [0.377046, 0.677222],
-  [0.344108, 0.679849],
-  [0.312876, 0.677668],
-  [0.283526, 0.66681],
-  [0.241246, 0.617214],
-  [0.102986, 0.531237],
-  [0.267612, 0.57544],
-  [0.297879, 0.566824],
-  [0.333434, 0.566122],
-  [0.366427, 0.573884],
-  [0.396012, 0.583304],
-  [0.420121, 0.589772],
-  [0.007561, 0.519223],
-  [0.432949, 0.430482],
-  [0.458639, 0.520911],
-  [0.473466, 0.454256],
-  [0.476088, 0.43617],
-  [0.468472, 0.444943],
-  [0.433991, 0.417638],
-  [0.483518, 0.437016],
-  [0.482483, 0.422151],
-  [0.42645, 0.610201],
-  [0.438999, 0.603505],
-  [0.450067, 0.599566],
-  [0.289712, 0.631747],
-  [0.27667, 0.636627],
-  [0.517862, 0.528052],
-  [0.710288, 0.619236],
-  [0.526227, 0.42609],
-  [0.895093, 0.745859],
-  [0.63407, 0.590424],
-  [0.661242, 0.586975],
-  [0.68888, 0.59054],
-  [0.725342, 0.610869],
-  [0.60663, 0.596295],
-  [0.654766, 0.655989],
-  [0.629906, 0.653924],
-  [0.680678, 0.652735],
-  [0.702097, 0.646409],
-  [0.752212, 0.589195],
-  [0.602918, 0.157137],
-  [0.719902, 0.6244],
-  [0.893693, 0.60004],
-  [0.790082, 0.608646],
-  [0.643998, 0.465512],
-  [0.528249, 0.349596],
-  [0.52585, 0.319809],
-  [0.560215, 0.342771],
-  [0.585384, 0.333459],
-  [0.549626, 0.319139],
-  [0.571228, 0.317308],
-  [0.624852, 0.271901],
-  [0.51305, 0.452718],
-  [0.515097, 0.472748],
-  [0.742247, 0.685493],
-  [0.598631, 0.545021],
-  [0.570338, 0.451425],
-  [0.578632, 0.466377],
-  [0.723087, 0.467946],
-  [0.516446, 0.500361],
-  [0.662801, 0.717082],
-  [0.703624, 0.706729],
-  [0.830705, 0.806186],
-  [0.552386, 0.697432],
-  [0.60761, 0.646112],
-  [0.645429, 0.303293],
-  [0.932695, 0.269895],
-  [0.557261, 0.427174],
-  [0.542902, 0.415208],
-  [0.618026, 0.305289],
-  [0.607591, 0.305797],
-  [0.722943, 0.728037],
-  [0.577414, 0.436833],
-  [0.614083, 0.718613],
-  [0.616907, 0.744114],
-  [0.668509, 0.880086],
-  [0.770092, 0.767979],
-  [0.635536, 0.810751],
-  [0.770391, 0.700444],
-  [0.826722, 0.721245],
-  [0.527121, 0.333802],
-  [0.553172, 0.331473],
-  [0.577238, 0.32611],
-  [0.554692, 0.419934],
-  [0.611897, 0.306039],
-  [0.596961, 0.29346],
-  [0.596371, 0.306047],
-  [0.539958, 0.442861],
-  [0.568842, 0.307634],
-  [0.547818, 0.307634],
-  [0.524613, 0.307634],
-  [0.53409, 0.220859],
-  [0.527671, 0.263774],
-  [0.526913, 0.282143],
-  [0.526878, 0.295374],
-  [0.526967, 0.304722],
-  [0.572058, 0.304722],
-  [0.573521, 0.29646],
-  [0.576838, 0.288154],
-  [0.581691, 0.279937],
-  [0.609945, 0.36009],
-  [0.986046, 0.439966],
-  [0.5868, 0.3046],
-  [0.590372, 0.298177],
-  [0.531915, 0.398463],
-  [0.577268, 0.414065],
-  [0.536915, 0.406214],
-  [0.627543, 0.526648],
-  [0.665586, 0.504049],
-  [0.588354, 0.453138],
-  [0.757824, 0.852324],
-  [0.70925, 0.798492],
-  [0.672684, 0.743419],
-  [0.600409, 0.250995],
-  [0.558266, 0.738328],
-  [0.570304, 0.812129],
-  [0.588166, 0.890956],
-  [0.711045, 0.601048],
-  [0.78107, 0.564595],
-  [0.587247, 0.601068],
-  [0.74287, 0.644554],
-  [0.572156, 0.562348],
-  [0.551868, 0.46343],
-  [0.821442, 0.542444],
-  [0.752702, 0.542818],
-  [0.713757, 0.532373],
-  [0.667113, 0.539327],
-  [0.631101, 0.552846],
-  [0.600862, 0.567527],
-  [0.523481, 0.594373],
-  [0.810748, 0.476074],
-  [0.771046, 0.651041],
-  [0.509127, 0.437282],
-  [0.595293, 0.514976],
-  [0.980531, 0.598436],
-  [0.5735, 0.58],
-  [0.602995, 0.451312],
-  [0.73353, 0.623023],
-  [0.560611, 0.480983],
-  [0.967686, 0.355643],
-  [0.580985, 0.61284],
-  [0.537728, 0.494615],
-  [0.760966, 0.220247],
-  [0.801779, 0.168062],
-  [0.892441, 0.459239],
-  [0.816351, 0.25974],
-  [0.865595, 0.666313],
-  [0.614074, 0.116754],
-  [0.508953, 0.420562],
-  [0.617942, 0.491684],
-  [0.825608, 0.602325],
-  [0.681215, 0.603765],
-  [0.656636, 0.599403],
-  [0.6039, 0.289783],
-  [0.812086, 0.411461],
-  [0.568013, 0.055435],
-  [0.681008, 0.101715],
-  [0.733752, 0.130299],
-  [0.63383, 0.601178],
-  [0.606793, 0.604463],
-  [0.58966, 0.608938],
-  [0.805016, 0.657892],
-  [0.611335, 0.637716],
-  [0.634038, 0.644029],
-  [0.656636, 0.644643],
-  [0.681215, 0.64166],
-  [0.698585, 0.636844],
-  [0.941867, 0.680924],
-  [0.698585, 0.612551],
-  [0.584177, 0.375893],
-  [0.554318, 0.433923],
-  [0.534154, 0.37936],
-  [0.711218, 0.180025],
-  [0.66463, 0.147129],
-  [0.5591, 0.097368],
-  [0.871706, 0.208059],
-  [0.591234, 0.626106],
-  [0.544341, 0.548416],
-  [0.624563, 0.075808],
-  [0.88577, 0.384971],
-  [0.551338, 0.304722],
-  [0.55198, 0.295368],
-  [0.552888, 0.284192],
-  [0.555168, 0.269206],
-  [0.569944, 0.232965],
-  [0.593203, 0.314324],
-  [0.599262, 0.318931],
-  [0.6076, 0.322297],
-  [0.631938, 0.3365],
-  [0.752033, 0.398685],
-  [0.547226, 0.579605],
-  [0.563544, 0.640172],
-  [0.583841, 0.631286],
-  [0.586614, 0.307634],
-  [0.771915, 0.316422],
-  [0.531597, 0.647517],
-  [0.588371, 0.195559],
-  [0.520797, 0.557435],
-  [0.567985, 0.506521],
-  [0.543283, 0.180745],
-  [0.655317, 0.254485],
-  [0.621009, 0.425982],
-  [0.62556, 0.219688],
-  [0.680198, 0.429281],
-  [0.642764, 0.395662],
-  [0.704663, 0.37847],
-  [0.552012, 0.137408],
-  [0.589072, 0.491363],
-  [0.685945, 0.224643],
-  [0.645735, 0.18736],
-  [0.675343, 0.296022],
-  [0.810858, 0.353695],
-  [0.720122, 0.285333],
-  [0.866152, 0.317295],
-  [0.663187, 0.355403],
-  [0.570082, 0.533674],
-  [0.544562, 0.451624],
-  [0.562759, 0.441215],
-  [0.531987, 0.46986],
-  [0.585271, 0.664823],
-  [0.622953, 0.677221],
-  [0.655896, 0.679837],
-  [0.687132, 0.677654],
-  [0.716482, 0.666799],
-  [0.758757, 0.617213],
-  [0.897013, 0.531231],
-  [0.732392, 0.575453],
-  [0.702114, 0.566837],
-  [0.666525, 0.566134],
-  [0.633505, 0.573912],
-  [0.603876, 0.583413],
-  [0.579658, 0.590055],
-  [0.99244, 0.519223],
-  [0.567192, 0.43058],
-  [0.541366, 0.521101],
-  [0.526564, 0.453882],
-  [0.523913, 0.43617],
-  [0.531529, 0.444943],
-  [0.566036, 0.417671],
-  [0.516311, 0.436946],
-  [0.517472, 0.422123],
-  [0.573595, 0.610193],
-  [0.560698, 0.604668],
-  [0.549756, 0.600249],
-  [0.710288, 0.631747],
-  [0.72333, 0.636627],
-];
-
-export default uvs;
