+++ src/face-target/aframe/face-occluder.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { AR_COMPONENT_NAME, AR_EVENT_NAME } from '..//utils/constant';
+import { AR_COMPONENT_NAME, AR_EVENT_NAME } from '../utils/constant';
 import { AR_ELEMENT_TAG } from '../../utils/constant';
 
 AFRAME.registerComponent(AR_COMPONENT_NAME.OCCULDER, {
