+++ src/face-target/face-geometry/face-geometry.ts
@@ -1,8 +1,8 @@
 import { uvs, faces } from './face-data';
 
-const nLandmarks = uvs.length;
+const N_LANDMARKS = uvs.length;
 
-const _createThreeFaceGeometry = (THREE: typeof AFRAME.THREE) => {
+const createThreeFaceGeometry = (THREE: typeof AFRAME.THREE) => {
   class FaceGeometry extends THREE.BufferGeometry {
     private positions: Float32Array;
     private uvs: Float32Array;
@@ -10,30 +10,35 @@ const _createThreeFaceGeometry = (THREE: typeof AFRAME.THREE) => {
     constructor() {
       super();
 
-      this.positions = new Float32Array(nLandmarks * 3);
-      this.uvs = new Float32Array(nLandmarks * 2);
+      this.positions = new Float32Array(N_LANDMARKS * 3);
+      this.uvs = new Float32Array(N_LANDMARKS * 2);
+
       this.setAttribute('position', new THREE.BufferAttribute(this.positions, 3));
       this.setAttribute('uv', new THREE.BufferAttribute(this.uvs, 2));
+
       this.setUvs();
+
       this.setIndex(faces);
     }
 
     setUvs() {
-      for (let j = 0; j < nLandmarks; j++) {
+      for (let j = 0; j < N_LANDMARKS; j++) {
         this.uvs[j * 2] = uvs[j][0];
         this.uvs[j * 2 + 1] = uvs[j][1];
       }
+
       this.getAttribute('uv').needsUpdate = true;
     }
 
     updatePositions(landmarks: number[][]) {
-      for (let i = 0; i < nLandmarks; i++) {
+      for (let i = 0; i < N_LANDMARKS; i++) {
         this.positions[i * 3 + 0] = landmarks[i][0];
         this.positions[i * 3 + 1] = landmarks[i][1];
         this.positions[i * 3 + 2] = landmarks[i][2];
       }
 
       this.attributes.position.needsUpdate = true;
+
       this.computeVertexNormals();
     }
   }
@@ -41,4 +46,4 @@ const _createThreeFaceGeometry = (THREE: typeof AFRAME.THREE) => {
   return new FaceGeometry();
 };
 
-export { _createThreeFaceGeometry };
+export default createThreeFaceGeometry;
