+++ src/face-target/face-mesh-helper.ts
@@ -1,8 +1,8 @@
 import { FaceMesh, Results } from '@mediapipe/face_mesh';
 
 class FaceMeshHelper {
-  private detectResolve: ((value: Results | PromiseLike<Results>) => void) | null;
   private faceMesh: FaceMesh;
+  private detectResolve: ((value: Results | PromiseLike<Results>) => void) | null;
 
   constructor() {
     this.detectResolve = null;
@@ -15,7 +15,6 @@ class FaceMeshHelper {
 
     this.faceMesh.setOptions({
       maxNumFaces: 1,
-      //refineLandmarks: true,
       refineLandmarks: false,
       minDetectionConfidence: 0.5,
       minTrackingConfidence: 0.5,
@@ -26,9 +25,10 @@ class FaceMeshHelper {
     });
   }
 
-  async detect(input: HTMLVideoElement) {
+  public async detect(input: HTMLVideoElement) {
     const results = await new Promise<Results>((resolve) => {
       this.detectResolve = resolve;
+
       this.faceMesh.send({ image: input });
     });
 
@@ -36,4 +36,4 @@ class FaceMeshHelper {
   }
 }
 
-export { FaceMeshHelper };
+export default FaceMeshHelper;
