+++ src/face-target/index.ts
@@ -1,4 +1,4 @@
-import { Controller } from './controller';
+import Controller from './controller';
 import { UI } from '../ui/ui';
 
 const faceTarget = {
